Adding the RIBs code generation template
Notable changes made:
- Added copyright to classes
- Removed previous internal references to classes outside RIBs
- Replaced ViewController and View references to UIViewController and
UIView references
- Modified the default RIB template to not have references to View /
Presenter. This is to make it be in line with Android. Yi: can you
please double check that this makes sense. Previously iOS did not have
a notion of a RIB with no Presenter & View, it only had a RIB with a
view, that did not own that view. (I might not have enough context here)
- Added an installation script to copy the template to the XCode
templates directory
- Added a basic readme