Add NEW_TASK_REPLACE RouterNavigator flag
This flag is similar to the NEW_TASK flag, but it always detaches all
existing routers and attaches the new one.

Currently, the DEFAULT flag is the only available flag that detaches the
top router if `newStateIsTop` is true. `NEW_TASK_REPLACE` combines this
property of `DEFAULT` with the behavior of `NEW_TASK`.