Clear `LazyCoroutineScope` cache upon `CoroutineScope` `Job` completion.
The `CoroutineScope` instance will be cached only until it is completed.

This new behavior of clearing the cache upon completion effectively allows for re-launching coroutines on the attach callback when an attach-detach-attach cycle occurs on `ScopeProvider` components with a lifecycle.

Re-launching coroutines upon reattachment is not possible in current implementation (without this fix): a single `CoroutineScope` is cached for the whole life of the instance, and `ScopeProvider.coroutineScope` delivers a stale, cancelled instance  of `CoroutineScope` after reattaching the component (even though `ScopeProvider.requestScope()` delivers a new `Completable`.

To be clear, the correct reattachment behavior is already in place for Rx subscriptions that `autoDispose` by `ScopeProvider`: they are backed by the `Completable`, which work fine on attach-detach-reattach cycles. This commit makes coroutines behavior consistent with the currently established behavior in RIBs.

Fixes #535.