Reduce Rx <-> Coroutines Interop and change `WorkerBinder.bind` logic.
This commit aims to minimize the amount of Rx <-> Coroutines interop to minimize the chances of surprising behavior. A flow converted to Rx converted to a Flow converted to Rx (...) will run multiple coroutines using Dispatchers.Undispatched (as per Flow.asObservable()), which will form event loops that we want to avoid.

This PR also changes worker binding implementation (bind(Interactor, Worker)) to use the non-converted Interactor lifecycle, and to launch binding coroutine in unconfined dispatcher while still starting undispatched.

Finally, it introduces a test that passes without f15ec67, fails with it, and passes again after this commit.